11nov/backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
Pillow>=10.0.0
joblib>=1.3.0
spotipy>=2.23.0
requests>=2.31.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
